requirements.txt
fastmcp
fastapi
uvicorn[standard]
pydantic
streamlit
requests
openai
python-dotenv
